refactor(builtins): extract shared resolve_path helper
Extract duplicated `resolve_path` function from 6 builtin modules into
a shared helper in `builtins/mod.rs`. This function resolves relative
paths against the current working directory.

Previously duplicated in:
- archive.rs
- curl.rs
- fileops.rs
- inspect.rs
- ls.rs
- pipeline.rs

Also moved tests for resolve_path to the module where it's defined.
