+++ crates/bashkit/src/builtins/archive.rs
@@ -10,7 +10,7 @@ use flate2::Compression;
 use std::io::{Read, Write};
 use std::path::Path;
 
-use super::{Builtin, Context};
+use super::{resolve_path, Builtin, Context};
 use crate::error::Result;
 use crate::interpreter::ExecResult;
 
@@ -774,16 +774,6 @@ impl Builtin for Gunzip {
         };
 
         Gzip.execute(new_ctx).await
-    }
-}
-
-/// Resolve a path relative to cwd
-fn resolve_path(cwd: &std::path::Path, path_str: &str) -> std::path::PathBuf {
-    let path = Path::new(path_str);
-    if path.is_absolute() {
-        path.to_path_buf()
-    } else {
-        cwd.join(path)
     }
 }
