+++ crates/bashkit/src/builtins/inspect.rs
@@ -1,9 +1,8 @@
 //! File inspection builtins - less, file, stat
 
 use async_trait::async_trait;
-use std::path::Path;
 
-use super::{Builtin, Context};
+use super::{resolve_path, Builtin, Context};
 use crate::error::Result;
 use crate::fs::FileType;
 use crate::interpreter::ExecResult;
@@ -396,16 +395,6 @@ fn default_stat_format(name: &str, metadata: &crate::fs::Metadata) -> String {
         access_str,
         modified,
     )
-}
-
-/// Resolve a path relative to cwd
-fn resolve_path(cwd: &std::path::Path, path_str: &str) -> std::path::PathBuf {
-    let path = Path::new(path_str);
-    if path.is_absolute() {
-        path.to_path_buf()
-    } else {
-        cwd.join(path)
-    }
 }
 
 #[cfg(test)]
