+++ crates/bashkit/src/builtins/pipeline.rs
@@ -1,9 +1,8 @@
 //! Pipeline control builtins - xargs, tee, watch
 
 use async_trait::async_trait;
-use std::path::Path;
 
-use super::{Builtin, Context};
+use super::{resolve_path, Builtin, Context};
 use crate::error::Result;
 use crate::interpreter::ExecResult;
 
@@ -264,16 +263,6 @@ impl Builtin for Watch {
         );
 
         Ok(ExecResult::ok(output))
-    }
-}
-
-/// Resolve a path relative to cwd
-fn resolve_path(cwd: &std::path::Path, path_str: &str) -> std::path::PathBuf {
-    let path = Path::new(path_str);
-    if path.is_absolute() {
-        path.to_path_buf()
-    } else {
-        cwd.join(path)
     }
 }
